Separate PR builds from CI builds
