+++ .github/workflows/develop.yml
@@ -1,9 +1,8 @@
 name: Lint and Test
 on:
-  push:
+  pull_request:
     branches:
       - "main"
-
 
 jobs:
   lint:
@@ -73,34 +72,3 @@ jobs:
 
       - name: Test With Race
         run: ENVIRONMENT=cicd make twr
-
-
-  build-and-push:
-    runs-on: ubuntu-latest
-    needs:
-      - lint
-      - test
-      - test-with-race
-    steps:
-      - name: 'Checkout The code'
-        uses: actions/checkout@master
-
-      - name: 'Extract Branch'
-        shell: bash
-        run: echo "##[set-output name=branch;]$(echo ${GITHUB_REF#refs/heads/})"
-        id: extract_branch
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v2
-
-      - name: Login to DockerHub
-        uses: docker/login-action@v2
-        with:
-          username: ${{ secrets.DOCKER_USERNAME }}
-          password: ${{ secrets.DOCKER_PASSWORD }}
-
-      - name: Build and push
-        uses: docker/build-push-action@v4
-        with:
-          push: true
-          tags: lordrahl/subscriptions:${{ steps.extract_branch.outputs.branch }}-${{ github.run_id }}-${{ github.run_number }}
