+++ servers/mocks_test.go
@@ -2,6 +2,7 @@ package servers
 
 import (
 	"context"
+	"subscriptions/domains/users"
 
 	"subscriptions/domains/plans"
 	"subscriptions/domains/products"
@@ -84,4 +85,25 @@ func (ps *planServiceMock) FindOne(ctx context.Context, id string) (*plans.Subsc
 
 func (ps *planServiceMock) Create(ctx context.Context, v *plans.SubscriptionPlan) error {
 	return ps.CreateFunc(ctx, v)
+}
+
+type userMock struct {
+	CreateFunc      func(ctx context.Context, u *users.User) error
+	FindFunc        func(ctx context.Context, id string) (*users.User, error)
+	FindByEmailFunc func(ctx context.Context, email string) (*users.User, error)
+}
+
+// Create implements users.IUserService
+func (um *userMock) Create(ctx context.Context, u *users.User) error {
+	return um.CreateFunc(ctx, u)
+}
+
+// Find implements users.IUserService
+func (um *userMock) Find(ctx context.Context, id string) (*users.User, error) {
+	return um.FindFunc(ctx, id)
+}
+
+// FindByEmail implements users.IUserService
+func (um *userMock) FindByEmail(ctx context.Context, email string) (*users.User, error) {
+	return um.FindByEmailFunc(ctx, email)
 }
